chore: rename license folder to licensing
