+++ backend/ee/licensing/utils.py
